pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "molprop-toolkit"
version = "0.5.0"
description = "MolProp Toolkit: calculators, analyzers, reports, picklists, and series analytics for medicinal chemistry property tables"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Easy-Peasy.AI" }]
keywords = ["cheminformatics", "rdkit", "medchem", "admet", "cns-mpo"]

dependencies = [
  "pandas>=1.5",
  "numpy>=1.23",
]

[project.optional-dependencies]
# Note: RDKit is best installed via conda-forge. This extra is provided for pip-only environments.
rdkit = ["rdkit-pypi>=2022.9.5"]
dimorphite = ["dimorphite-dl>=1.3.2"]
report = ["matplotlib>=3.7"]
dev = ["pytest>=7.4", "ruff>=0.1.9"]

[project.scripts]
# Calculators
molprop-calc-v4 = "molprop_toolkit.cli:calc_v4"
molprop-calc-v5 = "molprop_toolkit.cli:calc_v5"

# Analysis + reporting
molprop-analyze = "molprop_toolkit.cli:analyze"
molprop-report = "molprop_toolkit.cli:report"
molprop-sketch = "molprop_toolkit.cli:sketch"
molprop-picklists = "molprop_toolkit.cli:picklists"
molprop-series = "molprop_toolkit.cli:series"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
# We intentionally include top-level script packages so console scripts can import them.
where = ["."]
include = ["molprop_toolkit*", "tools", "analyzers", "calculators", "tests"]

[tool.setuptools.package-data]
# Ensure the SA fragment score file and schema are present in wheels.
calculators = ["data/*.pkl.gz"]
molprop_toolkit = ["resources/schema.json"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 120
exclude = [
  "calculators",
  "analyzers",
  "docs",
  "examples",
  "reports",
  "series",
]

[tool.ruff.lint]
# Start conservatively: only fail CI on clear bugs (undefined names, etc.).
select = ["F"]

